Split core startup to single file
